refactor: optimize + refactor refData aggregation
